refactor:#265 client my-page 리팩토링
