+++ src/app/(client)/my-page/_components/profile-bar.tsx
