+++ src/app/(client)/my-page/_utils/calculate-macronutrient-ratio.util.ts
